fix(message-queue): give correct -p flag to podman-compose
To build images with the correct prefix. The prefix was changed by
a29412b22f88f9a2277113d931fce8b78e483557 to include the repository name
to make the full name make sense in the context Github Actions and image
repository.

The previous prefix just happened to be the default because of the
parent directory name message-queue.

The issue went un-noticed for a while because I still had both images in the
local registry and didn't make any changes that would need an image
update.

refs: #10 #59
broken-by: a29412b22f88f9a2277113d931fce8b78e483557