refactor: run telemetry inside its own project
The issue with just including the telemetry services inside
message-queue/compose.yaml is that we would be building e.g.
code-golf_message-queue_telemetry-postgres.

Therefore

1. we would be building the telemetry service images for every
   user (e.g. message-queue) and
2. running those as separate services and
3. if we would need to build another image on top of it, the
   Containerfile needs to refer to the image full name, which I am not
   sure how would work. There is an option to also build the images with
   -p telemetry and then build on top of those but then we would need to
   make sure those are up-to-date with the user's project. Another
   option is to use ARG but that would likely mean we'd have to
   re-define the telemetry services to pass the arg.

Anyway, even if there would be clean solution to point 3. and 1. was
acceptable, it's likely that we want to keep point 2. as user
requirement in order to be able to run entire code-golf and get all the
results into one telemetry system.

The downside of this approach is that it slightly complicates the makefiles
and its not clear how the networking will work yet.

refs: #10