feat(.github): optimize GHA once again
Ok, I think this is finally working right.

We now have a constant minimal overhead of cca 30s to run the analyze
(cca 10s) and then the rest (20s) for GHA to figure out all the other
jobs are skipped, which is ridiculous but ok.

We build only if changes are detected. Good.

We run tests if appropriate changes are detected. That is e.g.:
1. base-debian is touched, anything depending on it is rebuilt and tested.
2. cpp producer is touched, cpp producer is rebuilt and tested using
   service images from master or last build of current branch if it has
   been modified on current branch.
3. If telemetry-postgres is touched, producer is not rebuilt but is
   retested again using existing image based on the same logic.

Good.

We can not use concurrency because the whole thing is kinda an inductive
step. So ideally we would need to make sure the workflows run
**correctly** on every push and in correct order without overlaps. I
don't think there is anything we can do about that but it also won't be
too much of an issue in practice.

We also get badges. Good. But we have to trigger each of the runtime
workflows via a REST API. Not good. We are risking a silent breakage
here but lets see how it works out.

Care needs to be taken to confirm all the needed CI checks were indeed
ran. There is also a bit of a race between when main workflow finishes
and the dispatched workflows start. That can result into a brief moment
where the comment is marked as all-checks-passed but there is still
dispatched workflow in the queue. That's the cost of having badges.

There's definitely more that can be done like running the API calls with
retries and also checking the status of code of the curl in dispatch action.

There is an enormous amount of boilerplate. It could probably be
resolved by wrapping the data we pass around as json objects but I'm not
sure if it can be done in the yaml or an executable action is needed to
handle it.

The runtime workflows could also probably be handled as a single
workflow with language input but again, we'd lose badges. But perhaps we
could just call re-usable workflow for that and get both.

The actions also does a lot. It would probably be worthwhile to look
into having a separate software that would determine the build and
runtime dependencies from some metadata files in the repository.

We should also get rid of the tj-actions/changed-files dependency and do
it ourselves as its default behaviors are total brain damage and we have
to do a bunch ourselves anyway. Namely,
1. It uses a default fetch_depth of 25 instead of just --deepen -ing as
   needed to get the correct diff, so at the moment we just clone the
   entire repo.
2. If a push is forced it simply looks just one commit behind instead of
   figuring the correct merge base, which we do, because I kinda force
   push a lot and it seems quite trivial to do. As long as the remote
   repo is not pruned of the old HEAD. And its pretty important since
   alternative would be to defaulting to rebuilding everything.

Note the fact that tj-actions/changed-files can't handle it isn't that
bad. But the fact that it just pretends it can, is.

We would also need GHA to implement some kind of pseudo-job that would
show up as a job on the overview board but it would run on the same
runner, possibly in parallel using matrix strategy as well but again, on
the same runner.

That would enable us to split analyze job into smaller logical pieces
without paying the job overhead (cca 12s) which is way too much just to
run detect action for each image separately.

refs: #63