refactor(message-queue/python): decouple events from their data
I'm torn on this. From OOP point of view, the events should be
encapsulated in their own objects. On the hand it would be just a weird
duplication and inefficiency in wrapping objects just to unwrap them
later.

Anyway, the first motivation was to be able to publish events as flag
combinations. Specifically to emit `WorkerResult | LastResult` but that
complicates the publish a lot. So I'll just avoid it for now.

I guess I like the separation of the event from its data in here.
Although, that's mainly because of 1. the simplification of the
subscribers, and 2. the deduplication between Events and primitives.
That however still could be achieved by merging the primitives into
events and dispatching based e.__class__.

I guess I'll really know when I do this in c++/rust.

refs: #78